+++ src/common/providers/fsProvider.ts
@@ -22,7 +22,8 @@ export class FSProvider implements IConfigProvider {
       const lastUpdatedDate = (await fsp.stat(this.fsConfig.yamlSourceFilePath)).mtime;
       return lastUpdatedDate;
     } catch (error) {
-      throw new Error(error);
+      const message = error instanceof Error ? error.message : JSON.stringify(error);
+      throw new Error(message);
     }
   }