+++ src/pollManager.ts
@@ -1,62 +1,65 @@
 import { Logger } from '@map-colonies/js-logger';
-import { inject, singleton } from 'tsyringe';
+import { $ } from 'zx';
+import { container, inject, singleton } from 'tsyringe';
 import { Watcher } from './watcher';
-import { Readiness } from './probe/readiness';
 import { Services } from './common/constants';
-import { IConfigProvider, IPollConfig } from './common/interfaces';
+import { IConfig, IConfigProvider, IPollConfig } from './common/interfaces';
 
 @singleton()
 export class PollManager {
+  private readonly config: IConfig;
+  private readonly gracefulReloadMaxSeconds: number;
+  private readonly msToSeconds: number;
+  private readonly fifoFilePath: string;
+
   public constructor(
     @inject(Services.LOGGER) private readonly logger: Logger,
     @inject(Services.POLLCONFIG) private readonly pollCofig: IPollConfig,
     @inject(Services.CONFIGPROVIDER) private readonly configProvider: IConfigProvider,
-    private readonly watcher: Watcher,
-    private readonly readiness: Readiness
-  ) {}
+    private readonly watcher: Watcher
+  ) {
+    this.config = container.resolve(Services.CONFIG);
+    this.gracefulReloadMaxSeconds = this.config.get<number>('gracefulReloadMaxSeconds');
+    this.fifoFilePath = this.config.get('uwsgiFifoFilePath');
+    this.msToSeconds = 1000;
+  }
 
   public async poll(): Promise<void> {
     const frequencyTimeOutMS = this.pollCofig.timeout.frequencyMilliseconds;
-    const requestsKillSeconds = this.pollCofig.timeout.requestsKillSeconds;
-    const afterUpdateDelaySeconds = this.pollCofig.timeout.afterUpdateDelaySeconds;
 
     try {
-      this.logger.info(`polling attempt`);
+      this.logger.debug(`polling attempt`);
       if (!(await this.watcher.isUpdated())) {
-        this.logger.debug('changes detected!');
-        const readinessKillRndInt = this.getRandomInteger();
-        this.logger.info(`killing readiness in ${readinessKillRndInt} seconds`);
-        await this.delay(readinessKillRndInt);
-        this.readiness.kill();
-        this.logger.info(`killing existing requests in ${requestsKillSeconds}`);
-        await this.delay(requestsKillSeconds);
-        this.logger.info('updating configurations');
+        this.logger.info('changes detected, updating configurations');
         await this.configProvider.createOrUpdateConfigFile();
-        // make sure mapproxy start updating before restart readiness
-        await this.delay(afterUpdateDelaySeconds);
-        this.readiness.start();
+
+        const gracefulReloadRandomSeconds = Math.floor(Math.random() * this.gracefulReloadMaxSeconds);
+        this.logger.info(`killing worker by graceful reload in uwsgi app within: ${gracefulReloadRandomSeconds} seconds`);
+        await this.delay(gracefulReloadRandomSeconds);
+
+        await this.reloadApp();
+        this.logger.info(`reload request was sent, app will be reloaded`);
       } else {
         this.logger.debug('no changes detected');
       }
-    } catch (error) {
-      if (error instanceof Error) {
-        const stackMessage = error.stack !== undefined ? error.stack : '';
-        this.logger.error(`Error occurred during poll check: ${error.message}, stack: ${stackMessage}`);
+    } catch (err) {
+      let errMsg = '';
+      if (err instanceof Error) {
+        errMsg = err.message ? err.message : '';
       }
+      this.logger.error({ msg: `Error occurred during poll check: ${errMsg}`, err: err });
     }
     setTimeout(() => {
       void this.poll();
     }, frequencyTimeOutMS);
   }
 
-  public async delay(seconds: number): Promise<void> {
-    const msToSeconds = 1000;
-    await new Promise((resolve) => setTimeout(resolve, seconds * msToSeconds));
+  public async reloadApp(): Promise<void> {
+    this.logger.info(`send reload App Request`);
+    await $`echo r > ${this.fifoFilePath}`;
   }
 
-  public getRandomInteger(): number {
-    const maxRandomSeconds = this.pollCofig.timeout.readinessKillMaxRandomSeconds;
-    const randomInt = Math.floor(Math.random() * maxRandomSeconds);
-    return randomInt;
+  public async delay(seconds: number): Promise<void> {
+    await new Promise((resolve) => setTimeout(resolve, seconds * this.msToSeconds));
   }
 }