+++ CHANGELOG.md
@@ -2,6 +2,60 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.2.4](https://github.com/MapColonies/mapproxinator/compare/v1.2.2...v1.2.4) (2024-03-14)
+
+
+### Bug Fixes
+
+* gracefulDelay will be taken from global ([#20](https://github.com/MapColonies/mapproxinator/issues/20)) ([26d4b46](https://github.com/MapColonies/mapproxinator/commit/26d4b46d57dfe130ff03d41a8fa038620a43569b))
+
+### [1.2.3](https://github.com/MapColonies/mapproxinator/compare/v1.2.2...v1.2.3) (2024-03-14)
+
+
+### Bug Fixes
+
+* gracefulDelay will be taken from global ([#20](https://github.com/MapColonies/mapproxinator/issues/20)) ([26d4b46](https://github.com/MapColonies/mapproxinator/commit/26d4b46d57dfe130ff03d41a8fa038620a43569b))
+
+### [1.2.2](https://github.com/MapColonies/mapproxinator/compare/v1.2.1...v1.2.2) (2024-01-21)
+
+### [1.2.1](https://github.com/MapColonies/mapproxinator/compare/v1.2.0...v1.2.1) (2023-12-26)
+
+## [1.2.0](https://github.com/MapColonies/mapproxinator/compare/v1.1.1...v1.2.0) (2023-12-25)
+
+
+### Features
+
+* support random seconds to graceful reload (MAPCO-3760) ([#17](https://github.com/MapColonies/mapproxinator/issues/17)) ([53abcd5](https://github.com/MapColonies/mapproxinator/commit/53abcd5199f6e1286203ff91517a1aa6208e439e))
+
+### [1.1.1](https://github.com/MapColonies/mapproxinator/compare/v1.1.0...v1.1.1) (2023-12-17)
+
+
+### Bug Fixes
+
+* handling db connection errors ([#16](https://github.com/MapColonies/mapproxinator/issues/16)) ([d8800c7](https://github.com/MapColonies/mapproxinator/commit/d8800c74f80ae65dd34c9be4554b7293396df9ff))
+
+## [1.1.0](https://github.com/MapColonies/mapproxinator/compare/v1.0.10...v1.1.0) (2023-12-12)
+
+
+### Features
+
+* replacing redness restart to graceful reload with fifo ([#15](https://github.com/MapColonies/mapproxinator/issues/15)) ([9d29ba3](https://github.com/MapColonies/mapproxinator/commit/9d29ba393b34296b57de9ee5e4a66ca721b5c764))
+
+### [1.0.10](https://github.com/MapColonies/mapproxinator/compare/v1.0.9...v1.0.10) (2023-10-10)
+
+
+### Bug Fixes
+
+* throw error if failed to provide config files (MAPCO-3567) ([#14](https://github.com/MapColonies/mapproxinator/issues/14)) ([12be070](https://github.com/MapColonies/mapproxinator/commit/12be070a36b02eb8d156c9d7dbab7c2269b4c96c))
+
+### [1.0.9](https://github.com/MapColonies/mapproxinator/compare/v1.0.8...v1.0.9) (2023-05-29)
+
+### [1.0.8](https://github.com/MapColonies/mapproxinator/compare/v1.0.5...v1.0.8) (2023-05-29)
+
+### [1.0.7](https://github.com/MapColonies/mapproxinator/compare/v1.0.5...v1.0.7) (2023-05-29)
+
+### [1.0.6](https://github.com/MapColonies/mapproxinator/compare/v1.0.5...v1.0.6) (2023-05-28)
+
 ### [1.0.5](https://github.com/MapColonies/mapproxinator/compare/v1.0.4...v1.0.5) (2022-12-27)
 
 ### [1.0.4](https://github.com/MapColonies/mapproxinator/compare/v1.0.3...v1.0.4) (2022-12-27)